feat: images pane
