+++ src/components/organisms_new/ExplorerPane/AccordionPanel/AccordionPanel.styled.tsx
