+++ src/components/organisms_new/ExplorerPane/FilePane/CreateNode.styled.tsx
