+++ src/components/organisms_new/ExplorerPane/FilePane/FilePane.styled.tsx
