+++ src/components/organisms_new/ExplorerPane/FilePane/FilePaneTree.styled.tsx
