+++ src/components/organisms_new/ExplorerPane/FilePane/TreeItem.styled.tsx
