--- src/components/organisms_new/ExplorerPane/HelmConfigPane/HelmConfigPane.tsx
@@ -1,11 +1,0 @@
-import {TitleBar} from '@monokle/components';
-
-const HelmConfigPane = () => {
-  return (
-    <div>
-      <TitleBar expandable isOpen title="Helm Configs" />
-    </div>
-  );
-};
-
-export default HelmConfigPane;
