--- src/components/organisms_new/ExplorerPane/HelmConfigPane/index.ts
@@ -1,1 +1,0 @@
-export {default} from './HelmConfigPane';
