+++ src/components/organisms_new/ExplorerPane/HelmPane/HelmPane.styled.tsx
