+++ src/components/organisms_new/ExplorerPane/HelmPane/HelmPane.tsx
@@ -1,10 +1,45 @@
-import {TitleBar} from '@monokle/components';
+import {CollapsePanelProps} from 'antd';
 
-const HelmPane = () => {
+import {size} from 'lodash';
+
+import {useAppSelector} from '@redux/hooks';
+
+import {SectionRenderer} from '@molecules';
+
+import RootHelmChartsSectionBlueprint from '@src/navsections/HelmChartSectionBlueprint';
+
+import {TitleBar, TitleBarCount} from '@monokle/components';
+import {InjectedPanelProps} from '@shared/models/explorer';
+
+import AccordionPanel from '../AccordionPanel';
+import {AccordionTitleBarContainer} from '../AccordionPanel/AccordionTitleBarContainer';
+import * as S from './HelmPane.styled';
+
+const HelmPane: React.FC<InjectedPanelProps> = props => {
+  const {isActive, panelKey} = props;
+
+  const helmChartMap = useAppSelector(state => state.main.helmChartMap);
+
   return (
-    <div>
-      <TitleBar expandable isOpen title="Helm Configs" />
-    </div>
+    <AccordionPanel
+      {...props}
+      header={
+        <AccordionTitleBarContainer>
+          <TitleBar
+            title="Helm"
+            expandable
+            isOpen={Boolean(isActive)}
+            actions={<TitleBarCount count={size(helmChartMap)} isActive={Boolean(isActive)} />}
+          />
+        </AccordionTitleBarContainer>
+      }
+      showArrow={false}
+      key={panelKey as CollapsePanelProps['key']}
+    >
+      <S.List id="helm-sections-container">
+        <SectionRenderer sectionBlueprint={RootHelmChartsSectionBlueprint} level={0} isLastSection={false} />
+      </S.List>
+    </AccordionPanel>
   );
 };
