--- src/components/organisms_new/ExplorerPane/ImagePane/ImagePane.tsx
@@ -1,11 +1,0 @@
-import {TitleBar} from '@monokle/components';
-
-const ImagePane = () => {
-  return (
-    <div>
-      <TitleBar expandable isOpen title="Helm Configs" />
-    </div>
-  );
-};
-
-export default ImagePane;
