--- src/components/organisms_new/ExplorerPane/ImagePane/index.ts
@@ -1,1 +1,0 @@
-export {default} from './ImagePane';
