+++ src/components/organisms_new/ExplorerPane/ImagesPane/ImagesPane.styled.tsx
