+++ src/components/organisms_new/ExplorerPane/ImagesPane/ImagesPane.tsx
@@ -6,19 +6,19 @@ import {useAppSelector} from '@redux/hooks';
 
 import {SectionRenderer} from '@molecules';
 
-import RootHelmChartsSectionBlueprint from '@src/navsections/HelmChartSectionBlueprint';
+import ImagesSectionBlueprint from '@src/navsections/ImagesSectionBlueprint';
 
 import {TitleBar, TitleBarCount} from '@monokle/components';
 import {InjectedPanelProps} from '@shared/models/explorer';
 
 import AccordionPanel from '../AccordionPanel';
 import {AccordionTitleBarContainer} from '../AccordionPanel/AccordionTitleBarContainer';
-import * as S from './HelmPane.styled';
+import * as S from './ImagesPane.styled';
 
-const HelmPane: React.FC<InjectedPanelProps> = props => {
+const ImagesPane: React.FC<InjectedPanelProps> = props => {
   const {isActive, panelKey} = props;
 
-  const helmChartMap = useAppSelector(state => state.main.helmChartMap);
+  const imagesList = useAppSelector(state => state.main.imagesList);
 
   return (
     <AccordionPanel
@@ -26,21 +26,21 @@ const HelmPane: React.FC<InjectedPanelProps> = props => {
       header={
         <AccordionTitleBarContainer>
           <TitleBar
-            title="Helm"
+            title="Images"
             expandable
             isOpen={Boolean(isActive)}
-            actions={<TitleBarCount count={size(helmChartMap)} isActive={Boolean(isActive)} />}
+            actions={<TitleBarCount count={size(imagesList)} isActive={Boolean(isActive)} />}
           />
         </AccordionTitleBarContainer>
       }
       showArrow={false}
       key={panelKey as CollapsePanelProps['key']}
     >
-      <S.List id="helm-sections-container">
-        <SectionRenderer sectionBlueprint={RootHelmChartsSectionBlueprint} level={0} isLastSection={false} />
+      <S.List id="images-section-container">
+        <SectionRenderer sectionBlueprint={ImagesSectionBlueprint} level={0} isLastSection={false} />
       </S.List>
     </AccordionPanel>
   );
 };
 
-export default HelmPane;
+export default ImagesPane;
