+++ src/components/organisms_new/ExplorerPane/ImagesPane/index.ts
@@ -1,0 +1,1 @@
+export {default} from './ImagesPane';
