+++ src/components/organisms_new/ExplorerPane/KustomizePane/KustomizePane.styled.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const List = styled.ol`
+  height: 100%;
+  list-style-type: none;
+  padding: 0;
+  padding-bottom: 20px;
+  overflow-y: auto;
+  margin: 0;
+`;
