--- src/components/organisms_new/NewExplorerPane/HelmPane/index.ts
@@ -1,1 +1,0 @@
-export {default} from './HelmPane';
