--- src/components/organisms_new/NewExplorerPane/KustomizePane/index.ts
@@ -1,1 +1,0 @@
-export {default} from './KustomizePane';
