--- src/components/organisms_new/NewExplorerPane/index.ts
@@ -1,1 +1,0 @@
-export {default} from './NewExplorerPane';
