+++ src/components/organisms_new/PaneManager/activities.tsx
@@ -11,7 +11,7 @@ import {ActivityType, Icon} from '@monokle/components';
 import {LeftMenuBottomSelectionType, LeftMenuSelectionType} from '@shared/models/ui';
 
 import CompareSyncPane from '../CompareSyncPane';
-import NewExplorerPane from '../NewExplorerPane';
+import ExplorerPane from '../ExplorerPane';
 import SettingsPane from '../SettingsPane';
 import ValidationPane from '../ValidationPane';
 
@@ -21,7 +21,7 @@ export const activities: ActivityType<LeftMenuSelectionType>[] = [
     name: 'explorer',
     tooltip: <FileExplorerTabTooltip />,
     icon: () => <Icon name="document" />,
-    component: <NewExplorerPane />,
+    component: <ExplorerPane />,
     useBadge: () => undefined,
     isVisible: () => useAppSelector(state => Boolean(!state.ui.isInQuickClusterMode)),
   },
