+++ src/components/organisms_new/index.ts
@@ -2,7 +2,7 @@ export {default as CompareSyncPane} from './CompareSyncPane';
 export {default as CreateProjectModal} from './CreateProjectModal';
 export {default as LearnPage} from './LearnPage';
 export {default as NavigatorPane} from './NavigatorPane';
-export {default as NewExplorerPane} from './NewExplorerPane';
+export {default as ExplorerPane} from './ExplorerPane';
 export {default as NewProject} from './NewProject';
 export {default as PaneManager} from './PaneManager';
 export {default as ProblemPane} from './ProblemPane';
